requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
yfinance>=0.2.0
langchain-google-genai>=0.1.0
langchain-core>=0.1.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
exa-py>=1.0.0
nest-asyncio>=1.5.0
seaborn>=0.12.0
matplotlib>=3.5.0
plotly>=5.0.0
sentence-transformers>=2.2.0
httpx>=0.25.0